+++ docs/source/conf.py
@@ -77,13 +77,13 @@ intersphinx_mapping = {
 #   alias ->  (base-URL, prefix)
 _repo = "https://github.com/mit-ll-responsible-ai/hydra-zen/"
 extlinks = {
-    "commit": (_repo + "commit/%s", "commit "),
-    "gh-file": (_repo + "blob/master/%s", ""),
-    "gh-link": (_repo + "%s", ""),
-    "issue": (_repo + "issues/%s", "issue #"),
-    "pull": (_repo + "pull/%s", "pull request #"),
-    "plymi": ("https://www.pythonlikeyoumeanit.com/%s", ""),
-    "hydra": ("https://hydra.cc/docs/%s", ""),
+    "commit": (_repo + "commit/%s", "commit %s"),
+    "gh-file": (_repo + "blob/master/%s", "%s"),
+    "gh-link": (_repo + "%s", "%s"),
+    "issue": (_repo + "issues/%s", "issue #%s"),
+    "pull": (_repo + "pull/%s", "pull request #%s"),
+    "plymi": ("https://www.pythonlikeyoumeanit.com/%s", "%s"),
+    "hydra": ("https://hydra.cc/docs/%s", "%s"),
 }
 
 # -- Options for HTML output -------------------------------------------------