+++ src/hydra_zen/wrapper/_implementations.py
@@ -2,11 +2,16 @@
 # SPDX-License-Identifier: MIR
 # pyright: strict
 
+from collections import defaultdict, deque
 from inspect import Parameter, signature
 from typing import (
     Any,
     Callable,
+    DefaultDict,
+    Deque,
     Dict,
+    FrozenSet,
+    Generator,
     Generic,
     Iterable,
     List,
@@ -18,28 +23,48 @@ from typing import (
     Type,
     TypeVar,
     Union,
+    cast,
     overload,
 )
 
 import hydra
-from hydra.main import _UNSPECIFIED_  # type: ignore
+from hydra.core.config_store import ConfigStore
+from hydra.utils import instantiate
 from omegaconf import DictConfig, ListConfig, OmegaConf
-from typing_extensions import Literal, ParamSpec, TypeAlias, TypeGuard
+from typing_extensions import (
+    Final,
+    Literal,
+    ParamSpec,
+    Protocol,
+    TypeAlias,
+    TypedDict,
+    TypeGuard,
+)
 
-from hydra_zen import instantiate
+from hydra_zen import instantiate, just, make_custom_builds_fn
 from hydra_zen.errors import HydraZenValidationError
-from hydra_zen.typing._implementations import DataClass_
+from hydra_zen.structured_configs._utils import get_obj_path
+from hydra_zen.typing._implementations import (
+    DataClass_,
+    GroupName,
+    Node,
+    NodeName,
+    StoreEntry,
+)
 
 from .._compatibility import HYDRA_SUPPORTS_LIST_INSTANTIATION, SUPPORTS_VERSION_BASE
 from ..structured_configs._type_guards import is_dataclass
 from ..structured_configs._utils import safe_name
 
-__all__ = ["zen"]
+__all__ = ["zen", "store", "Zen"]
 
 
 R = TypeVar("R")
 P = ParamSpec("P")
+F = TypeVar("F")
 
+
+_UNSPECIFIED_: Any = object()
 
 if HYDRA_SUPPORTS_LIST_INSTANTIATION:
     _SUPPORTED_INSTANTIATION_TYPES: Tuple[Any, ...] = (dict, DictConfig, list, ListConfig)  # type: ignore
@@ -249,6 +274,7 @@ class Zen(Generic[P, R]):
             raise HydraZenValidationError(
                 f"`cfg._args_` must be a sequence type (e.g. a list), got {_args_}"
             )
+
         if num_pos_only and len(_args_) != num_pos_only:
             raise HydraZenValidationError(
                 f"{self.func} has {num_pos_only} positional-only arguments, but "
@@ -336,7 +362,7 @@ class Zen(Generic[P, R]):
         version_base: Optional[str] = _UNSPECIFIED_,
     ) -> Callable[[Any], Any]:
         """
-        Returns a Hydra-CLI compatible version of the wrapped function: `hydra.main(zen(func))`
+        Generates a Hydra-CLI for the wrapped function. Equivalent to `hydra.main(zen(func), [...])()`
 
         Parameters
         ----------
@@ -361,8 +387,14 @@ class Zen(Generic[P, R]):
             Equivalent to `hydra.main(zen(func), [...])()`
         """
 
-        kw = dict(config_path=config_path, config_name=config_name)
-        if SUPPORTS_VERSION_BASE:
+        kw = dict(config_name=config_name)
+
+        if config_path is not _UNSPECIFIED_:
+            kw["config_path"] = config_path
+
+        if (
+            SUPPORTS_VERSION_BASE and version_base is not _UNSPECIFIED_
+        ):  # pragma: no cover
             kw["version_base"] = version_base
 
         return hydra.main(**kw)(self)()
@@ -541,7 +573,7 @@ def zen(
     ...     return x, zen_cfg
     >>> f(dict(x=1, y="${x}"))
     (1, {'x': 1, 'y': 1})
-    
+
     **Including a pre-call function**
 
     Given that a zen-wrapped function will automatically extract and instantiate config
@@ -629,3 +661,827 @@ def zen(
         )
 
     return wrap
+
+
+fbuilds = make_custom_builds_fn(populate_full_signature=True)
+
+
+def default_to_config(
+    target: Union[
+        Callable[..., Any],
+        DataClass_,
+        List[Any],
+        Dict[Any, Any],
+        ListConfig,
+        DictConfig,
+    ],
+    **kw: Any,
+) -> Union[DataClass_, Type[DataClass_], ListConfig, DictConfig]:
+    """Creates a config that describes `target`.
+
+    This function is designed to selectively apply `hydra_zen.builds` or
+    `hydra_zen.just` in a way that permits maximum compatibility with common
+    inputs to `hydra_zen.ZenStore`. It behavior can be summarized based on the type of
+    `target`
+
+    - OmegaConf containers and dataclass *instances* are returned unchanged
+    - A dataclass type is processed as `builds(target, **kw, populate_full_signature=True, builds_bases=(target,))`
+    - Lists and dictionaries are processed by `hydra_zen.just`
+    - All other inputs are processed as `builds(target, **kw, populate_full_signature=True)`
+
+    Parameters
+    ----------
+    target : Callable[..., Any] | DataClass | Type[DataClass] | list | dict
+
+    **kw : Any
+        Keyword arguments to be passed to `builds`.
+
+    Returns
+    -------
+    target_config :  DataClass | Type[DataClass] | list | dict
+
+    Examples
+    --------
+    Lists and dictionaries
+
+    >>> from hydra_zen.wrapper import default_to_config
+    >>> default_to_config([1, {"z": 2+2j}])
+    [1, {'z': ConfigComplex(real=2.0, imag=2.0, _target_='builtins.complex')}]
+
+    Dataclass types
+
+    >>> from dataclasses import dataclass
+    >>>
+    >>> @dataclass
+    ... class A:
+    ...     x: int
+    ...     y: int
+
+    >>> Builds_A = default_to_config(A, y=22)
+    >>> Builds_A(x=1)
+    Builds_A(x=1, y=22, _target_='__main__.A')
+    >>> issubclass(Builds_A, A)
+    True
+
+    A function
+
+    >>> from hydra_zen import to_yaml
+    >>> def func(x: int, y: int): ...
+    >>> print(to_yaml(default_to_config(func)))
+    _target_: __main__.func
+    x: ???
+    'y': ???
+    """
+    if is_dataclass(target):
+        if isinstance(target, type):
+            if get_obj_path(target).startswith("types."):
+                # handles dataclasses returned by make_config()
+                return target
+            return fbuilds(target, **kw, builds_bases=(target,))
+        if kw:
+            raise ValueError(
+                "store(<dataclass-instance>, [...]) does not support specifying "
+                "keyword arguments"
+            )
+        return target
+
+    elif isinstance(target, (dict, list)):
+        # TODO: convert to OmegaConf containers?
+        return just(target)
+    elif isinstance(target, (DictConfig, ListConfig)):
+        return target
+    else:
+        t = cast(Callable[..., Any], target)
+        return fbuilds(t, **kw)
+
+
+class _HasName(Protocol):
+    __name__: str
+
+
+# TODO: Should we automatically snake-case?
+def get_name(target: _HasName) -> str:
+    name = getattr(target, "__name__", None)
+    if not isinstance(name, str):
+        raise TypeError(
+            f"Cannot infer config store entry name for {target}. It does not have a "
+            f"`__name__` attribute. Please manually specify `store({target}, "
+            f"name=<some name>, [...])`"
+        )
+    return name
+
+
+class _StoreCallSig(TypedDict):
+    """Arguments for ZenStore.__call__
+
+    This default dict enables us to easily update/merge the default arguments for a
+    specific ZenStore instance, in support of self-partialing behavior."""
+
+    name: Union[NodeName, Callable[[Any], NodeName]]
+    group: Union[GroupName, Callable[[Any], GroupName]]
+    package: Optional[Union[str, Callable[[Any], str]]]
+    provider: Optional[str]
+    __kw: Dict[str, Any]  # kwargs passed to to_config
+    to_config: Callable[[Any], Any]
+
+
+# TODO: make frozen dict
+defaults: Final = _StoreCallSig(
+    name=get_name,
+    group=None,
+    package=None,
+    provider=None,
+    to_config=default_to_config,
+    __kw={},
+)
+
+_DEFAULT_KEYS: Final[FrozenSet[str]] = frozenset(_StoreCallSig.__required_keys__ - {"__kw"})  # type: ignore
+
+
+class _Deferred:
+    __slots__ = ("to_config", "target", "kw")
+
+    def __init__(
+        self, to_config: Callable[[F], Node], target: F, kw: Dict[str, Any]
+    ) -> None:
+        self.to_config = to_config
+        self.target = target
+        self.kw = kw
+
+    def __call__(self) -> Any:
+        return self.to_config(self.target, **self.kw)
+
+
+def _resolve_node(entry: StoreEntry, copy: bool) -> StoreEntry:
+    """Given an entry, updates the entry so that its node is not deferred, and returns
+    the entry. This function is a passthrough for an entry whose node is not deferred"""
+    item = entry["node"]
+    if isinstance(item, _Deferred):
+        entry["node"] = item()
+
+    if copy:
+        entry = entry.copy()
+    return entry
+
+
+class ZenStore:
+    """An abstraction over Hydra's config store, which enables users to maintain
+    multiple, isolated store instances before populating Hydra's global config store.
+
+    `ZenStore` is also designed to consolidate the config-creation and storage process;
+    it can be used to decorate config-targets and dataclasses, enabling "inline" config
+    creation and storage patterns. This is also a "self-partialing" object, meaning a
+    store instance can overwrite its own default values. Please consult the
+    :ref:`examples <self-partial>` for more details.
+
+    `hydra_zen.store` is available as a pre-instantiated, globally-available store,
+    which is initialized as:
+
+    .. code-block:: python
+
+       store = ZenStore(
+           name="zen_store",
+           deferred_to_config=True,
+           deferred_hydra_store=True,
+       )
+
+    Examples
+    --------
+    **Basic usage**
+
+    Let's add a config to hydra-zen's pre-instantiated `ZenStore` instance. Each config
+    must have an associated name. Optionally, a group, package, and/or provider may be
+    specified for the entry as well.
+
+    >>> from hydra_zen import store
+    >>> _ = store({"lr": 0.01, "momentum": 0.9}, name="sgd", group="optim")
+    >>> _ = store({"lr": 0.001}, name="adam", group="optim")
+    zen_store
+    {'optim': ['sgd', 'adam']}
+
+    A store's entries are keyed by their `(group, name)` pair (the default group is
+    `None`).
+
+    >>> store["optim", "sgd"]  # (group, name) -> config node
+    {'lr': 0.01, 'momentum': 0.9}
+
+    By default, the stored config(s) will be "enqueued" for addition to Hydra's config
+    store, and `.add_to_hydra_store()` must be called to add the enqueued configs
+    to Hydra's central store.
+
+    >>> store.has_enqueued()
+    True
+    >>> store.add_to_hydra_store()  # adds all enqueued entries to Hydra's global store
+    >>> store.has_enqueued()
+    False
+
+    By default, attempting to overwrite an entry will result in an error.
+
+    >>> store({}, name="sgd", group="optim")  # same name and group as above
+    ValueError: (name=sgd group=optim): Hydra config store entry already exists.
+    Specify `overwrite_ok=True` to enable replacing config store entries
+
+    We can create a distinct store that has an independent internal repository of
+    configs.
+
+    >>> from hydra_zen import ZenStore
+    >>> new_store = ZenStore("new_store")
+    >>> _ = new_store([1, 2, 3], name="backbone")
+
+    >>> store
+    zen_store
+    {'optim': ['sgd', 'adam']}
+    >>> new_store
+    new_store
+    {None: ['backbone']}
+
+    **Auto-config capabilities**
+
+    The input to a store is processed by the store's `to_config` function prior to
+    creating the stored config node. By default this calls
+    `hydra_zen.wrapper.default_to_config`, which applies `hydra_zen.builds` or
+    `hydra_zen.just` to inputs based on their types.
+
+    For instance, consider the following function:
+
+    >>> def sum_it(a: int, b: int): return a + b
+
+    We can pass `sum_it` directly to our store to leverage auto-config and auto-naming
+    capabilities. Here, `builds(sum_it, a=1, b=2)` will be called under the hood by
+    `new_store` to create the config for `sum_it`.
+
+    >>> from hydra_zen import to_yaml
+    >>> def pp(x): print(to_yaml(x))  # for pretty printing configs
+
+    >>> store2 = ZenStore()
+    >>> _ = store2(sum_it, a=1, b=2)  # entry name defaults to `sum_it.__name__`
+    >>> config = store2[None, "sum_it"]
+    >>> pp(config)
+    _target_: __main__.sum_it
+    a: 1
+    b: 2
+
+    Refer to `hydra_zen.wrapper.default_to_config` for more details about the default
+    auto-config behaviors of `ZenStore`.
+
+    **Support for decorator patterns**
+
+    `ZenStore.__call__` is a pass-through and can be used as a decorator. Let's add two
+    store entries for `func` by decorating it.
+
+    >>> from hydra_zen import ZenStore, to_yaml
+    >>> store = ZenStore()
+
+    >>> @store(a=1, b=22, name="func1")
+    ... @store(a=-10, name="func2")
+    ... def func(a: int, b: int):
+    ...     return a - b
+
+    >>> func(10, 3)  # the decorated function is left unchanged
+    7
+    >>> pp(store[None, "func1"])
+    _target_: __main__.func
+    a: 1
+    b: 22
+    >>> pp(store[None, "func2"])
+    _target_: __main__.func
+    b: ???
+    a: -10
+
+    Note that, by default, the application of `to_config` via the store is deferred
+    until that entry is actually accessed. This defers the runtime cost of constructing
+    configs for the decorated function so that it need not be paid until necessary.
+
+    .. _self-partial:
+
+    **Customizable store defaults via 'self-partialing' patterns**
+
+    The default values for a store's `__call__` parameters – `group`, `to_config`, etc.
+    – can easily be customized. Simpy call the store with those new values and
+    without specifying an object to be stored. This will return a "mirrored" store
+    instance – with the same internal state as the original store – with updated
+    defaults.
+
+    For example, let's create a store where we want to store multiple configs under a
+    `'math'` group and under a `'functools'` group, respectively.
+
+    >>> import math
+    >>> import functools
+
+    >>> new_store = ZenStore()
+    >>> math_store = new_store(group="math")  # overwrites group default
+    >>> tool_store = new_store(group="functools")  # overwrites group default
+
+    >>> math_store(math.floor)  # equivalent to: `new_store(math.floor, group="math")`
+    >>> math_store(math.ceil)
+    >>> tool_store(functools.lru_cache)
+    >>> tool_store(functools.wraps)
+
+    `math_store` and `tool_store` both share the same internal state as `new_store`, but
+    have overwritten default values for the `group`. See that `new_store` has entries
+    under these corresponding groups:
+
+    >>> new_store
+    custom_store
+    {'math': ['floor', 'ceil'], 'functools': ['lru_cache', 'wraps']}
+
+    These "self-partialing" patterns can be chained indefinitely and can be used to set
+    partial defaults for the `to_config` function.
+
+    >>> profile_store = new_store(group="profile")
+    >>> schemaless = profile_store(schema="<none>")
+
+    >>> from dataclasses import dataclass
+    >>> @profile_store(name="admin", has_root=True)
+    >>> @schemaless(name="test_admin", has_root=True)
+    >>> @schemaless(name="test_user", has_root=False)
+    >>> @dataclass
+    >>> class Profile:
+    >>>     username: str
+    >>>     schema: str
+    >>>     has_root: bool
+
+    >>> pp(new_store["profile", "admin"])
+    username: ???
+    schema: ???
+    has_root: true
+    _target_: __main__.Profile
+
+    >>> pp(new_store["profile", "test_admin"])
+    username: ???
+    schema: <none>
+    has_root: true
+    _target_: __main__.Profile
+    """
+
+    __slots__ = (
+        "name",
+        "_internal_repo",
+        "_defaults",
+        "_queue",
+        "_deferred_to_config",
+        "_deferred_store",
+        "_overwrite_ok",
+    )
+
+    def __init__(
+        self,
+        name: Optional[str] = None,
+        *,
+        deferred_to_config: bool = True,
+        deferred_hydra_store: bool = True,
+        overwrite_ok: bool = False,
+    ) -> None:
+        """
+        Parameters
+        ----------
+        name : Optional[str]
+            The name for this store.
+
+        deferred_to_config : bool, optional (default=True)
+            If `True` (default), this store will a not apply `to_config` to the
+            target until that specific entry is accessed by the store.
+
+        deferred_hydra_store : bool, optional (default=True)
+            If `True` (default), this store will not add entries to Hydra's global
+            config store until `store.add_to_hydra_store` is called explicitly.
+
+        overwrite_ok : bool, optional (default=False)
+            If `False` (default), attempting to overwrite entries in this store and
+            trying to use this store to overwrite entries in Hydra's global store
+            will raise a `ValueError`.
+        """
+        if not isinstance(deferred_to_config, bool):  # type: ignore
+            raise TypeError(
+                f"deferred_to_config must be a bool, got {deferred_to_config}"
+            )
+
+        if not isinstance(overwrite_ok, bool):  # type: ignore
+            raise TypeError(f"overwrite_ok must be a bool, got {overwrite_ok}")
+
+        if not isinstance(deferred_hydra_store, bool):  # type: ignore
+            raise TypeError(
+                f"deferred_hydra_store must be a bool, got {deferred_hydra_store}"
+            )
+
+        self.name: str = "custom_store" if name is None else name
+
+        # The following attributes are mirrored across store instances that are
+        # created via the 'self-partialing' process
+        self._internal_repo: Dict[Tuple[GroupName, NodeName], StoreEntry] = {}
+        # Internal repo entries that have yet to be added to Hydra's config store
+        self._queue: Deque[StoreEntry] = deque([])
+
+        self._deferred_to_config = deferred_to_config
+        self._deferred_store = deferred_hydra_store
+        self._overwrite_ok = overwrite_ok
+        # Contains the current default arguments for `self.__call__`
+        self._defaults: _StoreCallSig = defaults.copy()
+
+    def __repr__(self) -> str:
+        # TODO: nicer repr?
+        groups_contents: DefaultDict[Optional[str], List[str]] = defaultdict(list)
+        for grp, name in self._internal_repo:
+            groups_contents[grp].append(name)
+
+        return f"{self.name}\n{repr(dict(groups_contents))}"
+
+    def __eq__(self, __o: object) -> bool:
+        """Returns `True` if two stores share identical internal repos and queues.
+
+        Examples
+        --------
+        >>> from hydra_zen import ZenStore
+        >>> store1 = ZenStore()
+        >>> store2 = ZenStore()
+        >>> store1_a = store1(group='a')
+        >>> _ = store1_a(dict(x=1), name="foo")
+
+        >>> store1 == store1_a
+        True
+        >>> store1 == store2
+        False
+        """
+        if not isinstance(__o, ZenStore):
+            return False
+        return __o._internal_repo is self._internal_repo and __o._queue is self._queue
+
+    # TODO: support *to_config_pos_args
+    @overload
+    def __call__(
+        self,
+        __target: F,
+        *,
+        name: Union[NodeName, Callable[[F], NodeName]] = ...,
+        group: Union[GroupName, Callable[[F], GroupName]] = ...,
+        package: Optional[Union[str, Callable[[F], str]]] = ...,
+        provider: Optional[str] = ...,
+        to_config: Callable[[F], Node] = default_to_config,
+        **to_config_kw: Any,
+    ) -> F:  # pragma: no cover
+        ...
+
+    # TODO: ZenStore -> Self when mypy adds support for Self
+    #       https://github.com/python/mypy/pull/11666
+    @overload
+    def __call__(
+        self,
+        __target: Literal[None] = None,
+        *,
+        name: Union[NodeName, Callable[[Any], NodeName]] = ...,
+        group: Union[GroupName, Callable[[Any], GroupName]] = ...,
+        package: Optional[Union[str, Callable[[Any], str]]] = ...,
+        provider: Optional[str] = ...,
+        to_config: Callable[[Any], Node] = ...,
+        **to_config_kw: Any,
+    ) -> "ZenStore":  # pragma: no cover
+        ...
+
+    def __call__(self, __target: Optional[F] = None, **kw: Any) -> Union[F, "ZenStore"]:
+        """__call__(target : Optional[T] = None, /, name: NodeName | Callable[[Any], NodeName]] = ..., group: GroupName | Callable[[T], GroupName]] = None, package: Optional[str | Callable[[T], str]]] | None], provider: Optional[str], to_config: Callable[[T], Node] = ..., **to_config_kw: Any) -> T | ZenStore
+
+        The interface to an initialized store. Can be used to store a config or to
+        :ref:`customize the default values <self-partial>` of the store.
+
+        Parameters
+        ----------
+        obj : Optional[T]
+            The object to be stored. This is a **positional-only** argument.
+
+            If `obj` is not specified, then the provided arguments are used to create a
+            mirrored store instance with updated default arguments.
+
+        name : NodeName | Callable[[T], NodeName]
+            The entry's name, or a callable that will be called as
+            `(obj) -> entry-name`. The default is `lambda obj: obj.__name__`.
+            Store entries are keyed off of `(group, name)`.
+
+        group : Optional[GroupName | Callable[[T], GroupName]]
+            The entry's group's name, or a callable that will be called as
+            `(obj) -> entry-group`. The default is `None`. Subgroups can be
+            specified using / within the group name.
+
+            Store entries are keyed off of `(group, name)`.
+
+        to_config : Callable[[T], Node] = default_to_config
+            Called on `obj` to produce the entry's "node" (the config).
+            Refer to  `hydra_zen.wrapper.default_to_config` for the default
+            behavior. Specify `lambda x: x` to have `obj` be stored directly
+            as the entry's node.
+
+            By default the call to `to_config` is deferred until the entry
+            is actually accessed by the store.
+
+        package : Optional[str | Callable[[Any], str]]
+            The entry's package. Default is `None`.
+
+        provider : Optional[str]
+            An optional provider name for the entry.
+
+        **to_config_kw : Any
+            Additional arguments that will be passed to `to_config`.
+
+        Returns
+        -------
+        T | ZenStore
+            If `obj` was specified, it is returned unchanged. Otherwise a new instance
+            of `ZenStore` is return, which mirrors the internal state of this store and
+            has updated default arguments.
+        """
+        if __target is None:
+            _s = type(self)(
+                self.name,
+                deferred_to_config=self._deferred_to_config,
+                deferred_hydra_store=self._deferred_store,
+                overwrite_ok=self._overwrite_ok,
+            )
+            _s._defaults = self._defaults.copy()
+
+            # Important: mirror internal state *by reference* to ensure `_s` and
+            # `self` remain in sync
+            _s._internal_repo = self._internal_repo
+            _s._queue = self._queue
+
+            new_defaults: _StoreCallSig = {k: kw[k] for k in _DEFAULT_KEYS if k in kw}  # type: ignore
+
+            new_defaults["__kw"] = {
+                **_s._defaults["__kw"],
+                **{k: kw[k] for k in set(kw) - _DEFAULT_KEYS},
+            }
+            _s._defaults.update(new_defaults)
+            return _s
+        else:
+            to_config = kw.get("to_config", self._defaults["to_config"])
+            name = kw.get("name", self._defaults["name"])
+            group = kw.get("group", self._defaults["group"])
+            package = kw.get("package", self._defaults["package"])
+            provider = kw.get("provider", self._defaults["provider"])
+
+            _name: NodeName = name(__target) if callable(name) else name
+            if not isinstance(_name, str):  # type: ignore
+                raise TypeError(f"`name` must be a string, got {_name}")
+            del name
+
+            _group: GroupName = group(__target) if callable(group) else group
+            if _group is not None and not isinstance(_group, str):  # type: ignore
+                raise TypeError(f"`group` must be a string or None, got {_group}")
+            del group
+
+            _pkg = package(__target) if callable(package) else package
+            if _pkg is not None and not isinstance(_pkg, str):
+                raise TypeError(f"`package` must be a string or None, got {_pkg}")
+            del package
+
+            merged_kw = {
+                **self._defaults["__kw"],
+                **{k: kw[k] for k in set(kw) - _DEFAULT_KEYS},
+            }
+
+            if self._deferred_to_config:
+                node = _Deferred(to_config, __target, merged_kw)
+            else:
+                node = to_config(__target, **merged_kw)
+
+            entry = StoreEntry(
+                name=_name,
+                group=_group,
+                package=_pkg,
+                provider=provider,
+                node=node,
+            )
+
+            if not self._overwrite_ok and (_group, _name) in self._internal_repo:
+                raise ValueError(
+                    f"(name={entry['name']} group={entry['group']}): "
+                    f"Store entry already exists. Use a store initialized "
+                    f"with `ZenStore(overwrite_ok=True)` to overwrite config store "
+                    f"entries."
+                )
+            self._internal_repo[_group, _name] = entry
+            self._queue.append(entry)
+
+            if not self._deferred_store:
+                self.add_to_hydra_store()
+            return __target
+
+    @property
+    def groups(self) -> Sequence[GroupName]:
+        """Returns a sorted list of the groups registered with this store"""
+        set_: Set[GroupName] = set(group for group, _ in self._internal_repo)
+        if None in set_:
+            set_.remove(None)
+            no_none = cast(Set[str], set_)
+            return [None] + sorted(no_none)
+        else:
+            no_none = cast(Set[str], set_)
+            return sorted(no_none)
+
+    def has_enqueued(self) -> bool:
+        """`True` if this store has entries that have not yet been added to
+        Hydra's config store.
+
+        Returns
+        -------
+        bool
+
+        Examples
+        --------
+        >>> from hydra_zen import ZenStore
+        >>> store = ZenStore(deferred_hydra_store=True)
+        >>> store.has_enqueued()
+        False
+
+        >>> store({"a": 1}, name)
+        >>> store.has_enqueued()
+        True
+
+        >>> store.add_to_hydra_store()
+        >>> store.has_enqueued()
+        False
+        """
+        return bool(self._queue)
+
+    def __bool__(self) -> bool:
+        """`True` if entries have been added to this store, regardless of whether or
+        not they have been added to Hydra's config store"""
+        return bool(self._internal_repo)
+
+    @overload
+    def __getitem__(self, key: Tuple[GroupName, NodeName]) -> Node:  # pragma: no cover
+        ...
+
+    @overload
+    def __getitem__(
+        self, key: GroupName
+    ) -> Dict[Tuple[GroupName, NodeName], Node]:  # pragma: no cover
+        ...
+
+    def __getitem__(self, key: Union[GroupName, Tuple[GroupName, NodeName]]) -> Node:
+        """Access a entry's config node by specifying `(group, name)`. Or, access a
+        mapping of `(group, name) -> node` for all nodes in a specified group,
+        including nodes within subgroups.
+
+        See Also
+        --------
+        ZenStore.get_entry
+
+        Examples
+        --------
+        >>> from hydra_zen import store
+        >>> store(dict(x=1), name="a", group="fruit")
+        >>> store(dict(x=2), name="b", group="fruit/apple")
+        >>> store(dict(x=3), name="c", group="fruit/apple")
+        >>> store(dict(x=4), name="d", group="fruit/orange")
+        >>> store(dict(x=5), name="e", group="veggie")
+
+        Accessing an individual entry's config node.
+
+        >>> store["fruit/apple", "b"]
+        {'x': 2}
+
+        Accessing all config nodes under the "fruit/apple" group
+
+        >>> store["fruit/apple"]
+        {('fruit/apple', 'b'): {'x': 2}, ('fruit/apple', 'c'): {'x': 3}}
+
+        Accessing all config nodes under the "fruit" group
+
+        >>> store["fruit"]
+        {('fruit', 'a'): {'x': 1},
+         ('fruit/apple', 'b'): {'x': 2},
+         ('fruit/apple', 'c'): {'x': 3},
+         ('fruit/orange', 'd'): {'x': 4}}
+        """
+        # store[group] ->
+        #  {(group, name): node1, (group, name2): node2, (group/subgroup, name3): node3}
+        #
+        # store[group, name] -> node
+        if isinstance(key, str) or key is None:
+            key_not_none = key is not None
+            key_w_ender = key + "/" if key is not None else "<ZEN_NEVER>"
+            return {
+                (group, name): _resolve_node(entry, copy=False)["node"]
+                for (group, name), entry in self._internal_repo.items()
+                if group == key
+                or (
+                    key_not_none and group is not None and group.startswith(key_w_ender)
+                )
+            }
+        return _resolve_node(self._internal_repo[key], copy=False)["node"]
+
+    def get_entry(self, group: GroupName, name: NodeName) -> StoreEntry:
+        """Access a store entry, which is a mapping that specifies the entry's
+        name, group, package, provider, and node.
+
+        Notes
+        -----
+        Mutating the returned mapping will not affect the store's internal entry.
+        Mutating node in the returned entry may have unintended consequences and
+        is not advised.
+
+        Examples
+        --------
+        >>> from hydra_zen import store, ZenStore
+        >>> store(dict(x=1), name="a", group="fruit")
+        >>> store.get_entry("fruit", "a")
+        {'name': 'a',
+         'group': 'fruit',
+         'package': None,
+         'provider': None,
+         'node': {'x': 1}}
+        """
+        return _resolve_node(self._internal_repo[(group, name)], copy=True)
+
+    def __contains__(self, key: Union[GroupName, Tuple[GroupName, NodeName]]) -> bool:
+        """Checks if group or (group, node-name) exists in zen-store."""
+        if key is None:
+            return any(k[0] is None for k in self._internal_repo)
+
+        if isinstance(key, str):
+            key_w_end: str = key + "/"
+            return any(
+                key == group or group.startswith(key_w_end)
+                for group, _ in self._internal_repo
+                if group is not None
+            )
+        return key in self._internal_repo
+
+    def __iter__(self) -> Generator[StoreEntry, None, None]:
+        """Yields all entries in this store.
+
+        Notes
+        -----
+        Mutating the returned mappings will not affect the store's internal entries.
+        Mutating node in an entry may have unintended consequences and is not advised.
+
+        Examples
+        --------
+        >>> from hydra_zen import store
+        >>> store(dict(x=1), name="a", group="fruit")
+        >>> store(dict(x=2), name="b", group="fruit/orange")
+        >>> store(dict(x=3), name="c", group="veggie")
+
+        >>> list(store)
+        [{'name': 'a',
+          'group': 'fruit',
+          'package': None,
+          'provider': None,
+          'node': {'x': 1}},
+         {'name': 'b',
+          'group': 'fruit/orange',
+          'package': None,
+          'provider': None,
+          'node': {'x': 2}},
+         {'name': 'c',
+          'group': 'veggie',
+          'package': None,
+          'provider': None,
+          'node': {'x': 3}}]
+        """
+        yield from (_resolve_node(v, copy=True) for v in self._internal_repo.values())
+
+    def add_to_hydra_store(self, overwrite_ok: Optional[bool] = None) -> None:
+        """Adds all of this store's enqueued entries to Hydra's global config store.
+
+        Parameters
+        ----------
+        overwrite_ok : Optional[bool]
+            If `False`, this method raises `ValueError` if an entry in Hydra's config
+            store will be overwritten. Defaults to the value of `overwrite_ok`
+            specified when initializing this store."""
+
+        while self._queue:
+            entry = _resolve_node(self._queue.popleft(), copy=False)
+            if (
+                overwrite_ok is False
+                or (overwrite_ok is None and not self._overwrite_ok)
+            ) and self._exists_in_hydra_store(name=entry["name"], group=entry["group"]):
+                raise ValueError(
+                    f"(name={entry['name']} group={entry['group']}): "
+                    f"Hydra config store entry already exists. Specify "
+                    f"`overwrite_ok=True` to enable replacing config store entries"
+                )
+            ConfigStore.instance().store(**entry)
+
+    def _exists_in_hydra_store(
+        self,
+        *,
+        name: NodeName,
+        group: GroupName,
+        hydra_store: ConfigStore = ConfigStore().instance(),
+    ) -> bool:
+        repo = hydra_store.repo
+
+        if group is not None:
+            for group_name in group.split("/"):
+                repo = repo.get(group_name)
+                if repo is None:
+                    return False
+        return name + ".yaml" in repo
+
+
+store: ZenStore = ZenStore(
+    name="zen_store",
+    deferred_to_config=True,
+    deferred_hydra_store=True,
+)